+++ src/i18next.exports.js
@@ -2,16 +2,6 @@
 // If we're not in CommonJS, add `i18n` to the
 // global object or to jquery.
 if (typeof module !== 'undefined' && module.exports) {
-    if (!$) {
-      try {
-        $ = require('jquery');
-      } catch(e) {
-        // just ignore
-      }
-    }
-    if ($) {
-        $.i18n = $.i18n || i18n;
-    }
     module.exports = i18n;
 } else {
     if ($) {